src/pyproject.toml
[tool.poetry]
name = "task_manager"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django = "=2.2.16"
psycopg2-binary = "~=2.8"
pymongo = "~=3.3"
djangorestframework = "=3.12.2"
django-rest-swagger = "=2.2.0"
django-filter = "=2.4.0"
djangorestframework-filters = "=1.0.0.dev2"
django-model-utils = "=4.0.0"
django-cleanup = "=5.1.0"
django-cors-headers = "=3.5.0"
django-constance = {version = "2.7.0", extras = ["database"]}
django-celery-beat = "2.1.0"
django-extensions = "=3.0.9"
django-environ = "=0.4.5"
django-structlog = "=1.6.3"
markdown = "^3.2.2"
labelgun = { git = "https://gitlab.usetech.ru/pub/labelgun.git", extras = ["logger"], tag = "0.1.8" }
djangorestframework_simplejwt = "^4.4.0"
uwsgi = "~=2.0"
pydantic = "=1.7.2"

[tool.poetry.dev-dependencies]
black = "=19.3b0"
mixer = "=6.1.3"
pytest = "=6.1.2"
pytest-django = "=4.1.0"
pytest-mock = "^3.1.1"
pytest-env = "=0.6.2"
freezegun = "^1.0.0"
responses = "^0.12.0"

[tool.black]
line-length = 102
target_version = ['py36']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | src
  | front_end_part
  | docs
  | __pycache__
  | data
  | media
  | migrations
  | tests
)/
'''

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
